# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/feedback": {
            "get": {
                "summary": "Get all feedback with pagination and filters",
                "operationId": "dc2d3a20e7bf609f07d946e5c674fbb4",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Filter by rating (1-5)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 5,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "has_reply",
                        "in": "query",
                        "description": "Filter by reply status (true/false)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feedback retrieved successfully"
                    }
                }
            },
            "post": {
                "summary": "Create new feedback",
                "operationId": "00cc0cc44651586507dc21fec307d425",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating",
                                    "komentar"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "komentar": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Feedback created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/feedback/{id}": {
            "get": {
                "summary": "Get specific feedback by ID",
                "operationId": "0d837fe2e9fd2f07905744d25c0f73cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Feedback ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feedback retrieved successfully"
                    },
                    "404": {
                        "description": "Feedback not found"
                    }
                }
            },
            "put": {
                "summary": "Update user's own feedback",
                "operationId": "45f45dc205a225c558d1e644f755f5c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Feedback ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "komentar": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Feedback updated successfully"
                    },
                    "403": {
                        "description": "Forbidden - Can only edit own feedback"
                    },
                    "404": {
                        "description": "Feedback not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete user's own feedback",
                "operationId": "a067718a32f076a3daf710f3be3e1e44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Feedback ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feedback deleted successfully"
                    },
                    "403": {
                        "description": "Forbidden - Can only delete own feedback"
                    },
                    "404": {
                        "description": "Feedback not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/feedback/my": {
            "get": {
                "summary": "Get current user's feedback",
                "operationId": "066a689ca203b7eb1ea819db4921b69e",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User feedback retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}